requirements.txt
flask
streamlit
requests
python-dotenv
pandas
fuzzywuzzy
python-levenshtein
intuit-oauth
python-quickbooks
openpyxl
xlrd
numpy
scikit-learn
rapidfuzz
celery
redis
pyyaml
openai
anthropic
psutil
